requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
absl-py==2.1.0
    # via
    #   chex
    #   optax
    #   orbax-checkpoint
accelerate==0.31.0
    # via -r requirements.in
certifi==2024.6.2
    # via requests
charset-normalizer==3.3.2
    # via requests
chex==0.1.86
    # via optax
colorama==0.4.6
    # via tqdm
contourpy==1.2.1
    # via matplotlib
customtkinter==5.2.2
    # via -r requirements.in
cycler==0.12.1
    # via matplotlib
darkdetect==0.8.0
    # via customtkinter
diffusers==0.29.2
    # via -r requirements.in
etils[epath,epy]==1.7.0
    # via
    #   optax
    #   orbax-checkpoint
filelock==3.15.4
    # via
    #   diffusers
    #   huggingface-hub
    #   torch
    #   transformers
flax==0.8.5
    # via -r requirements.in
fonttools==4.53.0
    # via matplotlib
fsspec==2024.6.1
    # via
    #   etils
    #   huggingface-hub
huggingface-hub==0.23.4
    # via
    #   accelerate
    #   diffusers
    #   tokenizers
    #   transformers
idna==3.7
    # via requests
importlib-metadata==8.0.0
    # via diffusers
importlib-resources==6.4.0
    # via etils
invisible-watermark==0.2.0
    # via -r requirements.in
jax==0.4.30
    # via
    #   -r requirements.in
    #   chex
    #   flax
    #   optax
    #   orbax-checkpoint
jaxlib==0.4.30
    # via
    #   chex
    #   jax
    #   optax
    #   orbax-checkpoint
jinja2==3.1.4
    # via torch
kiwisolver==1.4.5
    # via matplotlib
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
matplotlib==3.9.0
    # via -r requirements.in
mdurl==0.1.2
    # via markdown-it-py
ml-dtypes==0.4.0
    # via
    #   jax
    #   jaxlib
    #   tensorstore
mpmath==1.3.0
    # via sympy
msgpack==1.0.8
    # via
    #   flax
    #   orbax-checkpoint
nest-asyncio==1.6.0
    # via orbax-checkpoint
networkx==3.3
    # via torch
numpy==1.26.4
    # via
    #   accelerate
    #   chex
    #   contourpy
    #   diffusers
    #   flax
    #   invisible-watermark
    #   jax
    #   jaxlib
    #   matplotlib
    #   ml-dtypes
    #   opencv-python
    #   opt-einsum
    #   optax
    #   orbax-checkpoint
    #   pywavelets
    #   scipy
    #   tensorstore
    #   transformers
opencv-python==4.10.0.84
    # via invisible-watermark
opt-einsum==3.3.0
    # via jax
optax==0.2.3
    # via flax
orbax-checkpoint==0.5.22
    # via flax
packaging==24.1
    # via
    #   accelerate
    #   customtkinter
    #   huggingface-hub
    #   matplotlib
    #   transformers
pillow==10.3.0
    # via
    #   diffusers
    #   invisible-watermark
    #   matplotlib
protobuf==5.27.2
    # via orbax-checkpoint
psutil==6.0.0
    # via accelerate
pygments==2.18.0
    # via rich
pyparsing==3.1.2
    # via matplotlib
python-dateutil==2.9.0.post0
    # via matplotlib
pywavelets==1.6.0
    # via invisible-watermark
pyyaml==6.0.1
    # via
    #   accelerate
    #   flax
    #   huggingface-hub
    #   orbax-checkpoint
    #   transformers
regex==2024.5.15
    # via
    #   diffusers
    #   transformers
requests==2.32.3
    # via
    #   diffusers
    #   huggingface-hub
    #   transformers
rich==13.7.1
    # via flax
safetensors==0.4.3
    # via
    #   -r requirements.in
    #   accelerate
    #   diffusers
    #   transformers
scipy==1.14.0
    # via
    #   jax
    #   jaxlib
six==1.16.0
    # via python-dateutil
sympy==1.12.1
    # via torch
tensorstore==0.1.63
    # via
    #   flax
    #   orbax-checkpoint
tokenizers==0.19.1
    # via transformers
toolz==0.12.1
    # via chex
torch==2.0.1
    # via
    #   accelerate
    #   invisible-watermark
tqdm==4.66.4
    # via
    #   huggingface-hub
    #   transformers
transformers==4.42.1
    # via -r requirements.in
typing-extensions==4.12.2
    # via
    #   chex
    #   etils
    #   flax
    #   huggingface-hub
    #   orbax-checkpoint
    #   torch
urllib3==2.2.2
    # via requests
zipp==3.19.2
    # via
    #   etils
    #   importlib-metadata
